pyproject.toml
[project]
name = "temperatureanalysisandmonitoring"
version = "0.1.0"
description = ""
authors = [
    {name = "Данил Краснов",email = "daankrasnov@edu.hse.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "jupyterlab (>=4.5.0,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "poetry-core (>=2.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pandas-stubs (>=2.3.3.251201)",
    "plotly (>=6.5.0,<7.0.0)",
    "streamlit (>=1.52.1,<2.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
